pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rawview"
version = "0.1.0"
description = "Qt front-end and agent for Ghidra reverse engineering"
requires-python = ">=3.11"
license = { file = "LICENSE" }
dependencies = [
    "PySide6>=6.6.0",
    "certifi>=2024.0.0",
    "anthropic>=0.40.0",
    "py4j>=0.10.9.7",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = ["pyinstaller>=6.0", "pillow>=10.0"]

[project.scripts]
rawview = "rawview.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["rawview*"]

[tool.setuptools.package-data]
rawview = ["qt_ui/resources/*.png", "qt_ui/resources/*.ico"]
